refactor(win32): Extract ClipboardData into a separate module
